Fixed most. Copy/Upload still broke

+++ JavaApp/src/java/com/algonquincollege/waterbin/fs/tasks/Remove.java
@@ -5,6 +5,7 @@
  */
 package com.algonquincollege.waterbin.fs.tasks;
 
+import com.algonquincollege.javaApp.database.DBConnection;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Paths;
@@ -14,11 +15,13 @@ import java.nio.file.Paths;
  * @author Devon St. John
  */
 public class Remove extends FileSystemTask {
+    private final DBConnection db;
     private final String path;
     
     public Remove(String path){
         this.path = path;
         
+        db = new DBConnection();
         System.out.println("Remove Task: Launched to Aggregator");
     }
     
@@ -26,8 +29,17 @@ public class Remove extends FileSystemTask {
     public void run() {
         System.out.println("Remove Task: " + path + " | Going Live");
         try{
-            Files.delete(Paths.get(root, path));
-            success = true;
+            if(db.connect() != null){
+                boolean wasFolder = Files.isDirectory(Paths.get(root, path));
+
+                Files.delete(Paths.get(root, path));
+                success = true;
+                if(wasFolder){
+                    db.deleteFolder(path);
+                }else{
+                    db.deleteFile(path);
+                }
+            }
         } catch (IOException ex) {
             System.err.println("Remove Failed: " + path);
         }
